1-ая итерация по 15 спринту

+++ src/main/java/ru/practicum/shareit/item/storage/ItemStorage.java
@@ -7,15 +7,15 @@ import ru.practicum.shareit.item.model.Item;
 import java.util.Collection;
 
 public interface ItemStorage {
-    public Item createItem(NewItemRequest newItemRequest, Integer userId);
+    public Item createItem(NewItemRequest newItemRequest, Long userId);
 
-    public Item updateItem(Long itemId, UpdateItemPatchRequest updateItemRequest, Integer userId);
+    public Item updateItem(Long itemId, UpdateItemPatchRequest updateItemRequest, Long userId);
 
-    public void removeItem(Long itemId, Integer userId);
+    public void removeItem(Long itemId, Long userId);
 
-    public Collection<Item> getAll(Integer userId);
+    public Collection<Item> getAll(Long userId);
 
     public Item getItem(Long itemId);
 
-    public Collection<Item> searchItem(String textSearch, Integer userId);
+    public Collection<Item> searchItem(String textSearch, Long userId);
 }
